dataops-infra/infra/requirements.txt
aws-cdk.core
aws-cdk-lib==2.165.0
constructs==10.4.2